Finalize allocation engine and improve allocation workflow

+++ app/Models/Allocation.php
@@ -17,7 +17,7 @@ class Allocation extends Model
     ];
 
     protected $casts = [
-        'score' => 'integer',
+        'score' => 'float',
     ];
 
     public function allocationRun(): BelongsTo
